requirements.txt
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
python-multipart>=0.0.5,<0.1.0
numpy>=1.21.0,<2.0.0
Pillow>=8.3.1,<9.0.0
pydantic>=1.8.0,<2.0.0
scikit-learn>=0.24.2,<1.0.0
python-dotenv>=0.19.0,<0.20.0
requests
